Fix absolute path failure and update metadata.
CPANTESTERs spotted a problem with abs_path on a file that didn't exist.
Removed the code that set the absolute path, not really needed. I was
being overly careful.
Gabor had an article reminding everyone to update LICENSE and repository
fields in Makefile.PL, so I did that while I was in there.